refactor util sub-pkgs to util pkg
